Update 1_Solved_Problems
Whenever we mention the name of the array , we get its BASE ADDRESS.

Therefore, first time through the loop, the printf() should print the value of the base address.

1_The Problem is in the next statement a++?

since, C doesn't perform bound checking on an array, the only thing that it remembers about an array once declared is its base address.

And a++ attempts to change this base address, which c won't allow because if it does so, it would be unable to remember the beginning of the array. 

***IMPORTANT***
Anything,which can change is called "lvalue" in compiler's language.
**********************

Since, value of a cannot be changed through++, it flashes the erroe saying 'L value required'.